Ready for integration testing

+++ src/main/java/org/joe/jobpoertalapp/services/JobSeekerService.java
@@ -3,10 +3,16 @@ package org.joe.jobpoertalapp.services;
 import org.joe.jobpoertalapp.dtos.incoming.InApplicationDto;
 import org.joe.jobpoertalapp.dtos.outgoing.OutApplicationDto;
 import org.joe.jobpoertalapp.dtos.outgoing.OutJobDto;
+import org.joe.jobpoertalapp.entities.Application;
+import org.joe.jobpoertalapp.entities.User;
+import org.joe.jobpoertalapp.exceptions.AccessDeniedException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class JobSeekerService {
@@ -29,16 +35,34 @@ public class JobSeekerService {
         return jobService.getAvailableJobsByTitleAndLocation(title,location);
     }
 
-    public List<OutApplicationDto> getMyApplications(Long jobSeekerId) {
-        return applicationService.viewApplicationByJobSeekerId(jobSeekerId);
+    public List<OutApplicationDto> getMyApplications() {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        User user = (User) auth.getPrincipal();
+        Long userId = user.getId();
+        return applicationService.viewApplicationByJobSeekerId(userId);
     }
 
     public void deleteApplication(Long id){
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        User user = (User) auth.getPrincipal();
+        Long userId = user.getId();
+
+        Application application = applicationService.getApplicationById(id);
+
+        if(!Objects.equals(application.getJobSeeker().getId(),userId)){
+            throw new AccessDeniedException("Access denied");
+        }
+
         applicationService.deleteApplication(id);
     }
 
     public OutApplicationDto createApplication(InApplicationDto inApplicationDto){
-        return applicationService.createApplication(inApplicationDto);
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        User user = (User) auth.getPrincipal();
+        Long userId = user.getId();
+
+
+        return applicationService.createApplication(inApplicationDto.jobId(),userId);
     }
 
 }
